Inline all monad methods
Makes all monad methods "inline" functions.
As higher-order functions, the monad methods are prime candidates for inlining.
This will iteratively replace the monad method definition in-place during compilation, leaving behind an imperative list of instructions with few, if any, intermediate lambda classes to create.
Since this drastically cuts down on the number of Function objects to create, it consequently cuts down on garbage collection, which should improve performance.
Since the monad methods all have implementations of a single expression, inlining does not drastically increase the size of the resulting code.