WIP - add DSN model to lander, add subcontext
Adds the subContext extension to SparkInitContext, to support hierarchical models.
This allows you to bake a hierarchical structure into the name of every cell and task, implicitly, to avoid naming conflicts.
By baking it into the context, we make it transparent to the model itself, both decluttering the model and avoiding errors due to forgetting to carry a base path around.

Using this subcontext, we clean up the Data model for the lander example, and add a DSN model.