WIP - debugging compile errors
There's a serious flaw in my current design, captured in the way cells include a serializer.
That design made sense when a serializer captured polymorphic subtypes internally; i.e., the "activity" serializer encoded what activity types are available.
This is *not* how Kotlin serializers work. In Kotlin, a polymorphic serializer is just a placeholder, and it's the serializers module that encodes which implementations are available.

To remedy this without polluting every cell with a reference to that module, which would be nigh impossible to implement, I'd like to put the module in the incon/fincon system.
This, and report handling, are the two places where we fundamentally need this capability.

The next few commits will rewrite this system to include a SerializersModule in the report handler and incon/fincon interfaces' JSON implementations, by generalizing the interface methods on type.
This may also provide opportunities to handle reporting or incons without going all the way to serial representations, which could be more performant.