Improve report handling
Adds a CSV report handler, as well as utilities for processing reports outside the simulation.

This allows a simulation author to better decouple modeling and reporting.
The model may now report its results in whichever format(s) are most amenable to writing an efficient model,
and the runner code can manipulate those reports to suit its needs.

For example, the EarthOrbit example model reports position vectors for the Earth and Moon as single resources with a record type.
This makes sense for the model, as the entire vector is a single datum.
When saving this data to CSV, it's better to split each vector component out to a separate column.
This is done in the report handler, since it's an artifact of choosing to save to CSV, not a modeling concern.