Eliminate fincon history for immediate awaits
Adds a NoOp TaskStepResult.
Instead of immediately blending into the next step, Restarts now produce an intermediate NoOp.
When the first step in a restarted task is an Await, and that awaiting task is saved, we wind up saving the history of the root task (an empty list)
instead of saving the history for the step before (the full history of the prior iteration).