Clean up
Assorted code cleanup, mostly suppressing some warnings.
In particular, suppresses the "nothing to inline" warnings on monad classes.
Rather than carefully pick and choose which monad methods to inline, it makes sense to just inline all of them.
This is simple and easy, and all the methods are small enough that inlining doesn't bloat the code.