Add runStandardPlanSimulation
Adds runStandardPlanSimulation, a function which defines a standard / baseline way to run Pyre.
This include file-based inputs and outputs, parallel reporting of outputs, and the JSON Lines format for simulation reports.

This function can be used directly by clients, but more importantly, functions as a baseline they can copy and then modify to suit their needs.