Add file-based resource
Adds utilities for precomputed / cached resources, chief among them fileBackedResource.
This function accepts the default simulation output format (a JSON Lines file of ChannelizedReports), selects one channel from it, and replays that channel as a resource in this simulation.

This is intended to support "layered" simulation architectures, where parts of the "full" simulation are run once in advance, and those results are replayed and fed into later layers to save computation.