WIP - Debugging CSV reporting
Last CSV format required us to know the CSV columns in advance.
For the first simulation, we can capture these by watching the initial resource reports come in.
This breaks down for later simulations.

I tried carrying a list of registered resources around, but it was very awkward.
In particular, it breaks the flow of giving the report handler to the simulation.
I tried to support this with lateinits and callbacks, but it was very ugly.

Instead, I've pivoted to "events CSV" format - each line is one event, rather than one time.
The columns are fixed in advance: time, channel, and data.
This also lends itself well to sparse resource changes, as only the changing resources are reported at all.
As a bonus, it also supports non-resource events like the activities channel, making it suitable as a catch-all default report handler.

A quick performance test with the EarthOrbit model shows no significant change to performance.