WIP - Debugging sequencing demo
There's a weird issue around coroutines throwing null pointers...
I think something to do with scopes is messing up and the final suspension is being sent to the wrong coroutine context.

I took this as an opportunity to simplify the scopes by cutting out the non-"Spark" scopes, but this didn't fully solve the issue.