WIP - Refactor scopes, debugging SequencingDemo
I believe the combination of receivers and context parameters for managing scopes (TaskScope, InitScope, ResourceScope, etc.) was causing weird issues with choosing the CoroutineTask to report results to.
I standardized everything to use context parameters, which also cleans up the code quite a bit.
I also took the opportunity to refactor and consolidate some of the scope-related functions, which had gotten quite scattered around.

Additionally, there was a bug in the core simulator around tracking conditions. I was keying off of the AwaitingTask object, which I rebuild, instead of the stable Await step I re-use. That was fixed.

Finally, I'm working on debugging the SequencingDemo model. There's some glitchiness around the command dispatch I need to fix still.
It may be worth building a version of the SequenceEngine that doesn't abuse the simulator quite as much. The sequenceBehavior cache is sketchy.