WIP - Debugging SequenceEngine
Rewrites the way we deal with control flow in a sequence engine.
The new approach extracts a list of "control flow points" from the sequence, storing them alongside the loaded sequence.
These points are serializable, so they can be stored in a regular discrete resource. Later, we might consider caching these somehow to compress the fincon file.