WIP - Initial design for unit-awareness
Adds a unit system and unit-aware wrapper, as well as some methods for working with unit-aware discrete double resources (so-called QuantityResources).
I'm pretty happy with how this looks in the demo class. I still need to run the demo class and make sure all the conversions are stacking up correctly.

I also need to write a suite of unit tests for the unit awareness. If that's going to be a critical component in people's models, it needs to be rock-solid.