WIP - Building up testing utilities for models
Adds a runUnitTest utiltity function.
This runs a "test task" in the simulation, and ends the simulation when the test task completes.
This borrows from the scheduling code to write results to memory as it runs, instead of to disk, so tests may write assertions against the results.
More conveniently, test assertions can be written within the test task itself, to run live against the model as it simulates.

Additionally, adds a test of the data subsystem model within the scheduling example.
The point of this is to demonstrate how an individual subsystem can easily by isolated and executed in its own test, decoupled from the rest of the model.