Add withInterrupts
Adds a wrapper withInterrupts, which can be used to run a tasks while monitoring for "interrupts".
An interrupt here is an arbitrary boolean resource, and is handled by an arbitrary task.
Multiple distinct interrupts may be specified, and the appropriate handler will be dispatched should an interrupt occur.
When an interrupt occurs, execution of the nominal task halts.