Flesh out profile operations
Flesh out profile operations, based on the idea of using a dedicated simulation to process simulation results.

The base tools are Profile.asResource, which lets us "replay" a profile as a resource in a simulation, and compute, which runs a simulation and packages the segments from a resource in that simulation into a profile.
Together, these make it trivial to process simulation results using the tools we already have for producing derived resources.

For convenience, some basic operators like arithmetic and comparators, are written for profiles. All of these just use compute and defer to their resource equivalents.

Note that this is currently untested.