Debug scheduling example
Debugs the scheduling example, and does some initial performance measurements on it.
The main performance driver now is copying and running the full system to schedule turns with a desired end time.
If we can leverage the new Profile.asResource() capabilities to build a GNC subsystem simulation to do this scheduling,
that should improve performance dramatically.