WIP - debugging partial incons
In attempting to build a GNC subsystem model to do turn scheduling in the scheduling example, I ran into the problem of a partial / misaligned initial conditions file.
This commit solves the problem of misaligned / missing root tasks, by adding in root tasks that are missing in the incon, and by reverting to the root task when children fail to restore cleanly.

There's still an issue with activities. Activities in the full system can be preserved through the subsystem fincon and restored as a task, which then fails to run because the wrong type of model is used.