Finish subsystem scheduling
Completes the subsystem scheduler to the extent I can, by adding a GNC-specific JSON format.
This forces the activity loader to restart, because full-system activities in the fincon can no longer be loaded.
There are still slight differences in the scheduling, but I think these are basically unavoidable.

Given the effort and idiosyncrasies involved in subsystem scheduling, I don't know if I'd advertise this approach broadly.
We at least proved it's possible, though.