Add test for intensive non-yielding tasks
Adds a test for a task which does a lot of reading and writing to a single cell without yielding.
This puts stress on the task- and cell-handling code. Specifically, if that code is recursive, it tends to cause stack overflow.
Recursion in that code can be non-obvious, so testing for it like this is necessary.