Eagerly apply effects
Switch CellSet to eagerly apply effects.
This means the CellState always tracks the most up-to-date value directly on each branch.
In the common case of merging a single nontrivial branch into the trunk cell history, that branch's cell is simply adopted as the trunk cell.
In the less-common case of multiple nontrivial branches, we can fall back on merging concurrent branches' effects like usual.

In practice, we expect this to improve performance by applying effects only once in the most common cases.
This is especially true for "intensive tasks", which perform a lot of work without yielding back to the engine.
These tasks can build up a large branch with a large effect, wasting cycles and threatening a stack overflow due to the highly-composed effect.