Add named scopes to tasks
Incorporates task names into the contextName for coroutine tasks, and forces all foundation tasks to use coroutines.
This is an acceptable loss of generality, as in practice writing any task not as a coroutine is very tedious.

As a result of this change, we get some slightly better insight into the "real" call chain when running, in the form of a context with a name recording the task lineage.