Move non-yielding kernel task actions to BasicTaskActions
At kernel level, when a simulation task performs an action which doesn't yield back to the engine, instead of returning this action as a continuation step, call the action directly in a newly-defined interface, BasicTaskActions, given to the task.

By threading those calls through to the foundation level, we should be able to remove the "suspend" requirement from non-yielding CoroutineTask actions.