Don't schedule listeners for immediately-satisfied conditions
In the common case that conditions are satisfied immediately, there's no reason to schedule the listeners for that awaiting task.
The continuation of the awaiting task will be run in the next task batch, and that continuation should just cancel the listeners anyways.

Optimize this path by not adding the listeners, nor the listener cleanup code, if a condition is satisfied immediately.

This made a scheduling simulation ~8% faster.