Clean up cell sets by remembering only the value
Keep the non-value information about a cell in the cell handle (now renamed cell), and store the value alone in the CellSet.
This is designed to both clean up the type hierarchy by coalescing cells and cell handles, and to maybe improve performance slightly by removing the need to copy cell data classes when stepping.
In practice, the performance implications are minimal.