Replace onStartup tasks with reporting directly during init
This is a safer way to report initial conditions, because we don't allow an "unsafe" onStartup task to arbitrarily modify the model.
To get this to work though, we need to initialize all cells with their correct value, rather than restoring them later.
This requires some refactoring, but ultimately leads to cleaner code.