Track cell bookkeeping variables in cells
Instead of keeping maps keyed on cells (so-called "cell handles"), puts the bookkeeping variables for a cell in the cell itself.
This is primarily to optimize reading a cell, as it eliminates a map lookup from that pathway.
As a side effect, it eliminates several unchecked casts, which necessarily involve some runtime checks, in favor of a few null checks.

This results in a ~4% reduction in simulation time, a surprisingly small change.