WIP - Don't serialize conditions
JsonConditions was eagerly serializing each value to JsonElement in memory.
This is wasteful for scheduling, when we turn around and immediately deserialize the values to initialize a simulation copy.

This commit removes this eager serialization, storing the in-memory objects instead.
As a side effect, a Json (format specification) is not required when constructing a JsonConditions (now a misnomer), cleaning up the constructor and separation of concerns.
One downside of this is that JsonConditions now requires a more complicated custom serializer, which can do the dynamics serializer lookups.
Additionally, the type of each value must now be serialized to enable the lookups during deserialization.

Serializing the types could be avoided if we instead deserialized each value to an arbitrary JsonElement, deferring the conversion to a real value until the `provide` call supplies us with a concrete KType.
This hybrid approach is left for future work.

This commit breaks many tests, for reasons I don't yet know.