WIP - debugging JsonConditions serialization
Hit a blocker for incremental reporting - since each value of the list can be a different KType, and I don't have anywhere to put that type information, it's hard to get this to work.

Instead of kludging even more type information into this structure, I'm considering just going for the deferred-deserialization setup, where we deserialize values when providing them.

That would be both more performant (likely) and require saving less type info.