WIP - Added TaskHistory, removed incremental conditions
Removes the incremental reporting/providing capabilities of FinconCollector and InconProvider.
As a consequence, collapses the FinconCollectingContext into FinconCollector and InconProvidingContext into InconProvider.
Replaces incremental conditions with TaskHistory, a type which can do the same job but is specific to Task history reporting.
The entire TaskHistory, once constructed, can then be reported as a unit to FinconCollector and retrieved from InconProvider.

This gets us to compiling code, but not passing tests.

Fundamentally, the issue is that it's hard to reliably serialize which type a value is.
Instead of refining this further, I'm going to try extending the approach I took for TaskHistory.
This will involve creating an in-memory Collector and Provider, a from-Json Provider, and an overarching interface hierarchy and multiplexing serializer to manage them.